pt6-apparatusL3-project/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.7
django-active-link==0.2.1
django-allauth==0.63.6
idna==3.7
isodate==0.6.1
lxml==5.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python3-saml==1.16.0
requests==2.32.3
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
wheel==0.43.0
xmlsec==1.3.14
pytz=2024.1

